+++ packages/nest-auth-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ackplus/nest-auth-react
 
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ackplus/nest-auth-client@2.7.5
+
 ## 2.7.4
 
 ### Patch Changes